LS-1-2: Ending implementation of sequence persistence.

+++ src/main/java/br/edu/femass/daos/Dao.java
@@ -6,7 +6,7 @@ public interface Dao<T> {
     void save(T t) throws Exception;
     List<T> getAll() throws Exception;
     T getByCode(Integer code) throws Exception;
-    Integer getNextCode() throws Exception;
+    Long getNextCode() throws Exception;
     void update(T t) throws Exception;
-    void delete(Integer code) throws Exception;
+    void delete(Long code) throws Exception;
 }
